requirements.txt
torch
pandas
numpy
tensorflow
torchvision
tokenizer
datasets